msm: vidc: Add support for VIDC_VIDEO_RATE_CONTROL_TIMESTAMP_MODE
In certain use cases, the timestamps provided by clients are uneven and
unpredictable. The unpredictability of timestamps causes uneveness in
the bitrate of encoded content. _TIMESTAMP_MODE_IGNORE allows the client
to indicate that its timestamps aren't to be trusted for rate control. In
these cases the core is expected to use it's own internal clock for rate
control.

Signed-off-by: Deva Ramasubramanian <dramasub@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>